Add custom BookNotFoundException

+++ src/main/java/com/example/libraryapp/domain/book/BookService.java
@@ -4,11 +4,11 @@ import com.example.libraryapp.domain.book.dto.BookDto;
 import com.example.libraryapp.domain.book.dto.BookToSaveDto;
 import com.example.libraryapp.domain.book.mapper.BookDtoMapper;
 import com.example.libraryapp.domain.book.mapper.BookToSaveDtoMapper;
+import com.example.libraryapp.domain.exception.BookNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.stream.StreamSupport;
 
@@ -50,7 +50,7 @@ public class BookService {
     @Transactional
     public void updateBook(Long id, BookDto book) {
         Book bookToUpdate = bookRepository.findById(id)
-                .orElseThrow(NoSuchElementException::new);
+                .orElseThrow(BookNotFoundException::new);
 
         if (book != null) {
             if (book.getTitle() != null) bookToUpdate.setTitle(book.getTitle());
@@ -65,6 +65,9 @@ public class BookService {
     }
 
     public void deleteBookById(Long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new BookNotFoundException();
+        }
         bookRepository.deleteById(id);
     }
 }
